Move Brackets Problem
This is my solution to the move brackets problem. Given is a random bracket sequence with an even number of brackets, half of them left brackets and the other half right brackets. What is required is to convert this sequence to a regular bracket sequence, by moving a number of brackets to the beginning or the end. The required answer is the minimum number of moves to achieve this.